# app/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "pint-server",
        "contact": {},
        "license": {
            "name": "MIT"
        }
    },
    "paths": {
...
        "/v3/envelopes/{envelopeReference}/additional-documents/{documentChecksum}": {
            "put": {
                "description": "Transfer an additional document (supporting document or eBL visualisation) associated with an eBL envelope transfer.\n\nThe receiving platform validates:\n- Document was declared in the EnvelopeManifest.\n- SHA-256 checksum matches the URL parameter.\n- Document size matches the manifest.\n\n**Request Body Format:**\n\nThe request body is a base64-encoded string containing the document content.\nExample: `UmF3IGNvbnRlbnQgb2YgdGhlIGZpbGU...` (plain base64 string, no JSON structure).\nThe decoded content type is determined by the sending platform based on the media type\ndeclared in the EnvelopeManifest.\n\nIf the sending platform loses track of the transfer state for a document, it can safely\nretry the transfer by resending the same document.\n\nIf the sending platform loses track of which documents have not been received, it can call\nthe PUT /v3/envelopes/{envelopeReference} endpoint again to get the current state.\n\n**Success Response:**\n\n`204 No Content` - This is returned when the document is received successfully,\nor when the document has already been received.\n\n\n**Error Responses (signed):**\n\n`409 Conflict` - Checksum or size mismatch (INCD response code).\n\n`422 Unprocessable Entity` - Envelope rejected (BSIG/BENV response code).\n",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PINT"
                ],
                "summary": "Transfer additional documents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Envelope reference (UUID)",
                        "name": "envelopeReference",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "SHA-256 checksum of the document",
                        "name": "documentChecksum",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Base64-encoded document content (plain string, e.g., UmF3IGNvbnRlbnQ...)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Document received successfully"
                    },
                    "400": {
                        "description": "Malformed request",
                        "schema": {
                            "$ref": "#/definitions/pint.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Envelope or document not found",
                        "schema": {
                            "$ref": "#/definitions/pint.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Checksum/size mismatch (INCD)",
                        "schema": {
                            "$ref": "#/definitions/pint.SignedEnvelopeTransferFinishedResponse"
                        }
                    },
                    "422": {
                        "description": "Envelope rejected (BSIG/BENV)",
                        "schema": {
                            "$ref": "#/definitions/pint.SignedEnvelopeTransferFinishedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/pint.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pint.DetailedError": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "description": "error code used on the platform: 7000-7999 for technical errors, 8000-8999 for functional errors",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pint.ErrorCode"
                        }
                    ]
                },
                "errorCodeMessage": {
                    "type": "string"
                },
                "errorCodeText": {
                    "type": "string"
                },
                "jsonPath": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "pint.ErrorCode": {
            "type": "integer",
            "enum": [
                7001,
                7002,
                7003,
                7004,
                7005,
                7006,
                7007,
                7008,
                7009,
                7010,
                7011,
                8001,
                8002
            ],
            "x-enum-comments": {
                "ErrCodeInsufficientTrust": "Trust level below minimum required",
                "ErrCodeMalformedRequest": "JSON parsing or encoding failed"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "JSON parsing or encoding failed",
                "",
                "",
                "",
                "",
                "",
                "",
                "Trust level below minimum required"
            ],
            "x-enum-varnames": [
                "ErrCodeBadSignature",
                "ErrCodeBadCertificate",
                "ErrCodeBadChecksum",
                "ErrCodeInvalidEnvelope",
                "ErrCodeInternalError",
                "ErrCodeMalformedRequest",
                "ErrCodeKeyError",
                "ErrCodeRegistryError",
                "ErrCodeRateLimitExceeded",
                "ErrCodeRequestTooLarge",
                "ErrCodeNotFound",
                "ErrCodeUnknownParty",
                "ErrCodeInsufficientTrust"
            ]
        },
        "pint.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorDateTime": {
                    "description": "The DateTime corresponding to the error occurring",
                    "type": "string"
                },
                "errors": {
                    "description": "An array of errors providing more detail about the root cause",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pint.DetailedError"
                    }
                },
                "httpMethod": {
                    "description": "The HTTP method used to make the request e.g. GET, POST, etc",
                    "type": "string"
                },
                "providerCorrelationReference": {
                    "description": "A unique identifier to the HTTP request within the scope of the API provider",
                    "type": "string"
                },
                "requestUri": {
                    "description": "The URI that was requested",
                    "type": "string"
                },
                "statusCode": {
                    "description": "The HTTP status code returned",
                    "type": "integer"
                },
                "statusCodeMessage": {
                    "description": "A long description corresponding to the HTTP status code with additional information",
                    "type": "string"
                },
                "statusCodeText": {
                    "description": "A standard short description corresponding to the HTTP status code",
                    "type": "string"
                }
            }
        },
...
        "pint.SignedEnvelopeTransferFinishedResponse": {
            "type": "object",
            "properties": {
                "envelopeTransferFinishedResponseSignedContent": {
                    "description": "EnvelopeTransferFinishedResponseSignedContent is a JWS-signed response returned when\nan envelope transfer is accepted or rejected.",
                    "type": "string",
                    "example": "eyJhbGciOiJFZERTQSIsImtpZCI6IjQ0MzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkM"
                }
            }
        },
...
    },
...
}